+++ src/gr/aueb/cf/cf9/cha18/BankApp/DTO/AccountReadOnlyDTO.java
@@ -1,4 +1,6 @@
 package gr.aueb.cf.cf9.cha18.BankApp.DTO;
 
-public record AccountReadOnlyDTO(String iban, double balance) {
+import java.math.BigDecimal;
+
+public record AccountReadOnlyDTO(String iban, BigDecimal balance) {
 }
